Use safer data type
Fix all potential data type mismatches. One example for data overflow is

    TestCli/-S_2491665: testsuite_cli_test.go:176: stderr:
        earlyoom v1.4
        fatal: -S: SIGTERM value 2491665 exceeds limit -2033605

caused by

  tuple = parse_term_kill_tuple(optarg, m.SwapTotalKiB * 100 / 99);

+++ meminfo.h
@@ -8,12 +8,12 @@
 
 typedef struct {
     // Values from /proc/meminfo, in KiB or converted to MiB.
-    long MemTotalKiB;
-    int MemTotalMiB;
-    int MemAvailableMiB; // -1 means no data available
-    int SwapTotalMiB;
-    long SwapTotalKiB;
-    int SwapFreeMiB;
+    long long MemTotalKiB;
+    long long MemTotalMiB;
+    long long MemAvailableMiB; // -1 means no data available
+    long long SwapTotalMiB;
+    long long SwapTotalKiB;
+    long long SwapFreeMiB;
     // Calculated percentages
     int MemAvailablePercent; // percent of total memory that is available
     int SwapFreePercent; // percent of total swap that is free
@@ -23,7 +23,7 @@ struct procinfo {
     int pid;
     int uid;
     int badness;
-    unsigned long VmRSSkiB;
+    long long VmRSSkiB;
     char name[PATH_LEN];
 };
 
@@ -32,7 +32,7 @@ bool is_alive(int pid);
 void print_mem_stats(int (*out_func)(const char* fmt, ...), const meminfo_t m);
 int get_oom_score(int pid);
 int get_oom_score_adj(const int pid, int* out);
-long get_vm_rss_kib(int pid);
+long long get_vm_rss_kib(int pid);
 int get_comm(int pid, char* out, int outlen);
 int get_uid(int pid);
